backend /backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
slowapi